requirements.txt
Django>=3.0,<4.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary  # Use for PostgreSQL
django-cors-headers
django-redis